修改refresh,修改个别计算公式

+++ src/PO/PlayerTechMPO.java
@@ -38,7 +38,7 @@ public class PlayerTechMPO implements Serializable{
 	public int opponentOffensiveRebound;                  //对手进攻篮板
 	public int opponentDefensiveRebound;                //对手防守篮板
 	public int teamShotIn;                             //全队进球数
-	public int opponentOffensiveNum;                     //对手进攻次数
+	public double opponentOffensiveNum;                     //对手进攻次数
 	public int opponentTwoShot;                     //对手进攻两分球出手次数
 	public int teamShot;                          //全队出手次数
 	public int teamPenaltyShot;                   //全队罚球次数
